Refactor and the README file has been updated to provide clearer documentation regarding the project structure
