+++ packages/tse-backend/src/app.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { AppService } from './app.service';
 import { ConfigModule, ConfigService } from '@nestjs/config';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { TypeOrmConfigService } from './config/typeorm-config.service';
+import { TypeOrmConfigService } from './common/config/typeorm-config.service';
 
 @Module({
   imports: [
