+++ packages/tse-backend/src/main.ts
@@ -3,7 +3,7 @@ import { AppModule } from './app.module';
 import { Logger, VersioningType } from '@nestjs/common';
 import { DocumentBuilder, SwaggerModule } from '@nestjs/swagger';
 import { ConfigService } from '@nestjs/config';
-import {GlobalExceptionFilter} from "./filters/global-exception.filter";
+import {GlobalExceptionFilter} from "./common/filters/global-exception.filter";
 
 async function bootstrap() {
   const logger = new Logger('bootstrap');
